+++ src/Heartbeat/Sensor/DBConnection.php
@@ -37,12 +37,4 @@ class DBConnection extends Sensor
             return false;
         }
     }
-
-    /**
-     * @inheritDoc
-     */
-    protected function getStatusMessage(bool $status): string
-    {
-        return $status ? __d('Heartbeat', 'OK') : __d('Heartbeat', 'FAILED');
-    }
 }